fix: modify typo and code convention for review
